backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.20
